replies destroy

+++ app/Http/Controllers/Api/RepliesController.php
@@ -26,11 +26,13 @@ class RepliesController extends Controller
             ->setStatusCode(201);
     }
 
-    public function destroy(Reply $reply)
+    public function destroy(Topic $topic, Reply $reply)
     {
+        if ($reply->topic_id != $topic->id) {
+            return $this->response->errorBadRequest();
+        }
         $this->authorize('destroy', $reply);
         $reply->delete();
-
-        return redirect()->to($reply->topic->link())->with('success', '成功删除回复！');
+        return $this->response->noContent();
     }
 }
